+++ NEWS.md
@@ -1,7 +1,9 @@
 # DedooseR 0.0.3
+* 2025-10-03: Updated the set_saturation function so we can set by either code counts or transcript coverage
+* 2025-10-03: Updated the plot_counts function so we can plot by either code counts or transcript coverage
+* 2025-10-03: Updated the count_codes function to output transcript coverage
 
 # DedooseR 0.0.2
-* 2025-10-03: Updated the count_codes function to output transcript coverage
 * 2025-10-02: Updated the clean_data function to include an exportable codebook!
 
 # DedooseR 0.0.1